refactor: security package move to config below
