+++ src/main/java/site/archive/api/v1/ArchiveControllerV1.java
@@ -22,8 +22,8 @@ import site.archive.api.v1.dto.archive.ArchiveListResponseDto;
 import site.archive.domain.archive.ArchiveImageService;
 import site.archive.domain.archive.ArchiveService;
 import site.archive.domain.user.info.UserInfo;
-import site.archive.security.authz.annotation.RequirePermission;
-import site.archive.security.authz.ArchiveAdminOrAuthorChecker;
+import site.archive.config.security.authz.annotation.RequirePermission;
+import site.archive.config.security.authz.ArchiveAdminOrAuthorChecker;
 
 @RestController
 @RequestMapping("/api/v1/archive")
