+++ src/main/java/site/archive/api/v1/RegisterControllerV1.java
@@ -15,8 +15,8 @@ import site.archive.api.v1.dto.user.OAuthRegisterRequestDto;
 import site.archive.domain.user.UserRegisterService;
 import site.archive.domain.user.info.UserInfo;
 import site.archive.infra.user.oauth.OAuthUserService;
-import site.archive.security.token.HttpAuthTokenSupport;
-import site.archive.security.token.TokenProvider;
+import site.archive.config.security.token.HttpAuthTokenSupport;
+import site.archive.config.security.token.TokenProvider;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
