+++ src/main/java/site/archive/config/security/authz/ArchiveAdminOrAuthorChecker.java
@@ -1,10 +1,10 @@
-package site.archive.security.authz;
+package site.archive.config.security.authz;
 
 import lombok.RequiredArgsConstructor;
+import site.archive.config.security.authz.permissionhandler.ArchivePermissionHandler;
 import site.archive.domain.archive.ArchiveRepository;
 import site.archive.domain.user.entity.UserRole;
 import site.archive.domain.user.info.UserInfo;
-import site.archive.security.authz.permissionhandler.ArchivePermissionHandler;
 
 @RequiredArgsConstructor
 public class ArchiveAdminOrAuthorChecker implements ArchivePermissionHandler {
