+++ src/main/java/site/archive/config/security/common/handler/LoginSuccessHandler.java
@@ -1,14 +1,14 @@
-package site.archive.security.result;
+package site.archive.config.security.common.handler;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import site.archive.config.security.token.HttpAuthTokenSupport;
 import site.archive.domain.user.UserService;
-import site.archive.security.common.UserPrincipal;
-import site.archive.security.token.HttpAuthTokenSupport;
-import site.archive.security.token.TokenProvider;
+import site.archive.config.security.common.UserPrincipal;
+import site.archive.config.security.token.TokenProvider;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
