+++ src/main/java/site/archive/util/SecurityUtils.java
@@ -3,7 +3,7 @@ package site.archive.util;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import site.archive.domain.user.info.UserInfo;
-import site.archive.security.token.jwt.JwtAuthenticationToken;
+import site.archive.config.security.token.jwt.JwtAuthenticationToken;
 
 import java.security.SecureRandom;
